Moved logger to Contracts
