+++ Realmius.Contracts/Logger/ConsoleLogger.cs
@@ -1,0 +1,26 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Realmius.Contracts.Logger
+{
+    public class ConsoleLogger : ILogger
+    {
+        public void Exception(Exception ex, string text = null)
+        {
+            System.Diagnostics.Debug.WriteLine($"Exception: {ex}, {text}");
+        }
+
+        public void Info(string text)
+        {
+            System.Diagnostics.Debug.WriteLine(text);
+        }
+
+        public void Debug(string text)
+        {
+            System.Diagnostics.Debug.WriteLine(text);
+        }
+    }
+}
