+++ Realmius.Contracts/Logger/ILoggerAware.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Realmius.SyncService.ApiClient
+namespace Realmius.Contracts.Logger
 {
     public interface ILoggerAware
     {
