+++ Realmius.Server/Configurations/RealmiusConfigurationBase.cs
@@ -21,6 +21,7 @@ using System.Collections.Generic;
 using Microsoft.AspNet.SignalR;
 using Realmius.Server.Infrastructure;
 using Realmius.Server.Models;
+using Realmius.Contracts.Logger;
 
 namespace Realmius.Server.Configurations
 {
